Move all block related code into BlockTranslator It makes more sense.
